separate bar chart slice
