+++ packages/frontend/src/features/sqlRunner/store/index.ts
@@ -1,17 +1,15 @@
 // FIXES ts2742 issue with configureStore
 // eslint-disable-next-line @typescript-eslint/no-unused-vars, no-unused-vars
 
-import { combineReducers, configureStore } from '@reduxjs/toolkit';
+import { configureStore } from '@reduxjs/toolkit';
 import { barChartSlice } from './barChartSlice';
 import { sqlRunnerSlice } from './sqlRunnerSlice';
 
-const rootReducer = combineReducers({
-    sqlRunner: sqlRunnerSlice.reducer,
-    barChartSlice: barChartSlice.reducer,
-});
-
 export const store = configureStore({
-    reducer: rootReducer,
+    reducer: {
+        sqlRunner: sqlRunnerSlice.reducer,
+        barChartConfig: barChartSlice.reducer,
+    },
     devTools: process.env.NODE_ENV === 'development',
 });
