+++ packages/frontend/src/features/sqlRunner/store/sqlRunnerSlice.ts
@@ -1,6 +1,5 @@
 import {
     ChartKind,
-    type BarChartConfig,
     type ResultRow,
     type SqlChart,
     type SqlTableConfig,
@@ -26,7 +25,6 @@ export interface SqlRunnerState {
     selectedChartType: ChartKind;
 
     resultsTableConfig: SqlTableConfig | undefined;
-    barChartConfig: BarChartConfig | undefined;
     tableChartConfig: TableChartSqlConfig | undefined;
 
     modals: {
@@ -45,7 +43,6 @@ const initialState: SqlRunnerState = {
     activeVisTab: VisTabs.CHART,
     selectedChartType: ChartKind.VERTICAL_BAR,
     resultsTableConfig: undefined,
-    barChartConfig: undefined,
     tableChartConfig: undefined,
     modals: {
         saveChartModal: {
@@ -98,41 +95,6 @@ export const sqlRunnerSlice = createSlice({
                     columns,
                 };
             }
-
-            // TODO: this initialization should be put somewhere it
-            // can be shared between the frontend and backend
-            if (state.barChartConfig === undefined) {
-                const fieldIds = Object.keys(action.payload[0]);
-                state.barChartConfig = {
-                    metadata: {
-                        version: 1,
-                    },
-                    type: ChartKind.VERTICAL_BAR,
-                    style: {
-                        legend: {
-                            position: 'top',
-                            align: 'center',
-                        },
-                    },
-                    axes: {
-                        x: {
-                            reference: fieldIds[0],
-                            label: fieldIds[0],
-                        },
-                        y: [
-                            {
-                                reference: fieldIds[1],
-                                label: fieldIds[1],
-                            },
-                        ],
-                    },
-                    series: fieldIds.slice(1).map((reference, index) => ({
-                        reference,
-                        name: reference,
-                        yIndex: index,
-                    })),
-                };
-            }
         },
         setSql: (state, action: PayloadAction<string>) => {
             state.sql = action.payload;
@@ -150,8 +112,6 @@ export const sqlRunnerSlice = createSlice({
                     : ChartKind.VERTICAL_BAR;
             if (action.payload.config.type === ChartKind.TABLE) {
                 state.tableChartConfig = action.payload.config;
-            } else if (action.payload.config.type === ChartKind.VERTICAL_BAR) {
-                state.barChartConfig = action.payload.config;
             }
         },
         updateTableChartFieldConfigLabel: (
